old_project/requirements.txt
requests
pandas==2.1.0
python-dotenv==1.0.0
PyYAML
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4
google-cloud-bigquery-storage==2.32.0
streamlit==1.31.0
plotly==5.18.0